requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests
python-dotenv
mcp

